requirements.txt
clarifai==9.8.1
streamlit==1.24.0
python-dotenv
